+++ crates/bevy_winit/src/winit_windows.rs
@@ -39,7 +39,7 @@ impl WinitWindows {
     /// Creates a `winit` window and associates it with our entity.
     pub fn create_window(
         &mut self,
-        event_loop: &winit::event_loop::EventLoopWindowTarget<crate::UserEvent>,
+        event_loop: &winit::event_loop::EventLoopWindowTarget<crate::WakeUp>,
         entity: Entity,
         window: &Window,
         adapters: &mut AccessKitAdapters,
